Relocate utils and don't test pages
Relocate the utils file to its own directory instead of being in
`src/components/`.  Don't test the pages directory since we test
individual components on their own.
